supressed set-but-not-used-variable warning in options parser
The gengetopt program (which produces the command-line parser
for cptutils programs) seems to have lost its maintainer, and
so the long-standing (but harmless) bug which includes an unused
variable in the parser has not been fixed.  This gives a slew of
compiler warnings which mask real problems, so I've added a
special compile rule for the files options.o adding the compiler
option -Wno-set-but-not-used-variable to supress these.  Immedaitely
I found a minor bug in cptinfo!